move test to loader
