+++ crates/loader/src/lib.rs
@@ -94,3 +94,49 @@ fn single_file_manifest(
 
     Ok(manifest)
 }
+
+#[cfg(test)]
+mod test {
+    use std::collections::HashSet;
+
+    use spin_app::{retain_components, App};
+
+    use super::*;
+
+    pub async fn build_locked_app(manifest: &toml::Table) -> anyhow::Result<LockedApp> {
+        let toml_str = toml::to_string(manifest).context("failed serializing manifest")?;
+        let dir = tempfile::tempdir().context("failed creating tempdir")?;
+        let path = dir.path().join("spin.toml");
+        std::fs::write(&path, toml_str).context("failed writing manifest")?;
+        from_file(&path, FilesMountStrategy::Direct, None).await
+    }
+
+    fn does_nothing_validator(_: &App, _: &[&str]) -> anyhow::Result<()> {
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_retain_components_filtering_for_only_component_works() {
+        let manifest = toml::toml! {
+            spin_manifest_version = 2
+
+            [application]
+            name = "test-app"
+
+            [[trigger.test-trigger]]
+            component = "empty"
+
+            [component.empty]
+            source = "does-not-exist.wasm"
+        };
+        let mut locked_app = build_locked_app(&manifest).await.unwrap();
+        locked_app = retain_components(locked_app, &["empty"], &[&does_nothing_validator]).unwrap();
+        let components = locked_app
+            .components
+            .iter()
+            .map(|c| c.id.to_string())
+            .collect::<HashSet<_>>();
+        assert!(components.contains("empty"));
+        assert!(components.len() == 1);
+    }
+}
